requirements.txt
django
django-crispy-forms
mysqlclient
pillow
torch
opencv-python
pandas
requests
pyyaml
tqdm
torchvision
matplotlib
seaborn